Перенес сервисы и ресты контроллеров все что связанно с картинками в ImageController, Добавил проверку на комментарии
