+++ src/main/java/ru/skypro/homework/controller/CommentsController.java
@@ -2,6 +2,7 @@ package ru.skypro.homework.controller;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
 import ru.skypro.homework.dto.CommentDTO;
 import ru.skypro.homework.projections.Comments;
@@ -30,13 +31,14 @@ public class CommentsController {
     }
 
     @DeleteMapping("/{adId}/comments/{commentId}")
-    public void deleteComment(@PathVariable int adId, @PathVariable int commentId) {
-        commentsService.deleteComment(adId, commentId);
+    public void deleteComment(@PathVariable int adId, @PathVariable int commentId, Authentication authentication) {
+        commentsService.deleteComment(adId, commentId, authentication);
     }
 
     @PatchMapping("/{adId}/comments/{commentId}")
     public CommentDTO updateComment(@PathVariable int adId, @PathVariable int commentId,
-                                    @Valid @RequestBody CreateOrUpdateComment createOrUpdateComment) {
-        return commentsService.updateComment(adId, commentId, createOrUpdateComment);
+                                    @Valid @RequestBody CreateOrUpdateComment createOrUpdateComment,
+                                    Authentication authentication) {
+        return commentsService.updateComment(adId, commentId, createOrUpdateComment, authentication);
     }
 }
