+++ src/main/java/ru/skypro/homework/controller/UserController.java
@@ -1,11 +1,7 @@
 package ru.skypro.homework.controller;
 
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.MediaType;
-import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
-import org.springframework.web.multipart.MultipartFile;
 import ru.skypro.homework.dto.UserDTO;
 import ru.skypro.homework.projections.NewPassword;
 import ru.skypro.homework.projections.UpdateUser;
@@ -21,26 +17,29 @@ public class UserController {
 
     private final UserServiceImpl userService;
 
-    // Получение пользователя
+    /**
+     * Получение пользователя
+     */
     @GetMapping("/me")
     public UserDTO getUser() {
         return userService.getUser();
     }
 
+    /**
+     * Изменение пароля пользователя
+     */
     @PostMapping("/set_password")
     public void setPassword(@RequestBody @Valid NewPassword newPassword) {
         userService.updatePassword(newPassword);
     }
 
+    /**
+     * Изменение пользователя
+     */
     @PatchMapping("/me")
     public UpdateUser updateUser(@RequestBody UpdateUser updateUser) {
         return userService.updateUser(updateUser);
     }
 
-    @PatchMapping(value = "/me/image", consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
-    public ResponseEntity<?> updateUserImage(@RequestPart("image") MultipartFile image,
-                                             Authentication authentication) {
 
-        return ResponseEntity.ok(userService.updateImage(image, authentication));
-    }
 }
