+++ src/main/java/ru/skypro/homework/service/CommentsService.java
@@ -9,6 +9,7 @@ public interface CommentsService {
     Comments getComments(int id);
     Comments getAllComments(int adId);
     CommentDTO addComment(int id, CreateOrUpdateComment createOrUpdateComment);
-    void deleteComment(int id, int commentsId);
-    CommentDTO updateComment(int id, int commentsId, CreateOrUpdateComment createOrUpdateComment);
+    void deleteComment(int id, int commentsId, Authentication authentication);
+    CommentDTO updateComment(int id, int commentsId,
+                             CreateOrUpdateComment createOrUpdateComment, Authentication authentication);
 }
