+++ src/main/java/ru/skypro/homework/service/ImageService.java
@@ -1,7 +1,14 @@
 package ru.skypro.homework.service;
 
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
+import org.springframework.web.multipart.MultipartFile;
 
 public interface ImageService {
     ResponseEntity<?> getImage(String id);
+
+    String updateImage(int id, MultipartFile file);
+
+    String updateImage(MultipartFile file, Authentication authentication);
+
 }
