+++ src/main/java/ru/skypro/homework/service/UserService.java
@@ -1,12 +1,8 @@
 package ru.skypro.homework.service;
 
-import org.springframework.data.crossstore.ChangeSetPersister;
-import org.springframework.security.core.Authentication;
-import org.springframework.web.multipart.MultipartFile;
 import ru.skypro.homework.dto.UserDTO;
 import ru.skypro.homework.model.UserModel;
 import ru.skypro.homework.projections.NewPassword;
-import ru.skypro.homework.projections.Register;
 import ru.skypro.homework.projections.UpdateUser;
 
 import java.util.Optional;
@@ -21,5 +17,4 @@ public interface UserService {
 
     UpdateUser updateUser(UpdateUser updateUser);
 
-    String updateImage(MultipartFile file, Authentication authentication);
 }
