+++ src/main/java/ru/skypro/homework/service/impl/ImageServiceImpl.java
@@ -4,25 +4,89 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.core.io.InputStreamResource;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.multipart.MultipartFile;
+import ru.skypro.homework.exceptions.AdNotFoundException;
 import ru.skypro.homework.exceptions.ImageNotFoundException;
+import ru.skypro.homework.exceptions.UserNotFoundException;
+import ru.skypro.homework.model.AdModel;
 import ru.skypro.homework.model.ImageModel;
+import ru.skypro.homework.model.UserModel;
+import ru.skypro.homework.repository.AdRepo;
 import ru.skypro.homework.repository.ImageRepo;
+import ru.skypro.homework.repository.UserRepo;
 import ru.skypro.homework.service.ImageService;
 
 import java.io.ByteArrayInputStream;
+import java.io.IOException;
+import java.util.Objects;
+import java.util.UUID;
 
 @Service
 public class ImageServiceImpl implements ImageService {
     @Autowired
     ImageRepo imageRepo;
+    @Autowired
+    AdRepo adRepo;
+    @Autowired
+    UserRepo userRepo;
 
+    /**
+     * Отображение картинок на странице
+     */
     public ResponseEntity<?> getImage(String id) {
         ImageModel image = imageRepo.findById(id).orElseThrow(ImageNotFoundException::new);
         assert image != null;
         return ResponseEntity.ok()
                 .contentType(MediaType.IMAGE_JPEG)
                 .body(new InputStreamResource(new ByteArrayInputStream(image.getBytes())));
+    }
+
+    /**
+     * Изменение картинки объявления
+     */
+    @Transactional
+    @Override
+    public String updateImage(int id, MultipartFile file) {
+        AdModel ad = adRepo.findById(id).orElseThrow(AdNotFoundException::new);
+
+        ImageModel imageModel = imageRepo.findById(ad.getImage().getId()).orElseThrow(ImageNotFoundException::new);
+        try {
+            imageModel.setBytes(file.getBytes());
+        } catch (IOException e) {
+            throw new RuntimeException(e);
+        }
+        imageRepo.saveAndFlush(imageModel);
+        return ("/image/" + imageModel.getId());
+    }
+
+    /**
+     * Обновление аватара пользователя
+     */
+    @Transactional
+    @Override
+    public String updateImage(MultipartFile file, Authentication authentication) {
+        UserModel userModel = userRepo.findByUserName(authentication.getName()).orElseThrow(UserNotFoundException::new);
+        ImageModel imageModel;
+        if (!Objects.isNull(userModel.getImage())) {
+            imageModel = imageRepo.findById(userModel.getImage().getId()).orElse(new ImageModel());
+        } else {
+            imageModel = new ImageModel();
+            imageModel.setId(UUID.randomUUID().toString());
+        }
+        try {
+            byte[] imageBytes = file.getBytes();
+            imageModel.setBytes(imageBytes);
+        } catch (IOException e) {
+            throw new RuntimeException();
+        }
+
+        imageRepo.saveAndFlush(imageModel);
+        userModel.setImage(imageModel);
+        userRepo.save(userModel);
+        return ("/image/" + imageModel.getId());
     }
 
 }
